requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-MySQLdb==1.0.1
passlib==1.7.4
Flask-Mail>=0.9.1
Flask-JWT-Extended==4.3.1
requests==2.27.1
Flask-Cors==3.0.10
xlrd==2.0.1
pandas~=1.4.1
gunicorn==20.1.0
waitress~=2.1.1
PyJWT~=2.3.0
SQLAlchemy~=1.4.32
Werkzeug==2.2.2
